updated leet.py

+++ leet.py
@@ -20,9 +20,11 @@ os.makedirs(contest_name, exist_ok=True)
 os.chdir('{0}/{1}'.format(os.getcwd(), contest_name))
 os.makedirs(problem_name, exist_ok=True)
 os.chdir('{0}/{1}'.format(os.getcwd(), problem_name))
-# TODO: 1.cppの存在チェック(あればコピーしない)
-shutil.copy(
-    '{0}/_lib/cpp/oj-configs/tpl-leet.cpp'.format(project_root),
-    '{0}/1.cpp'.format(os.getcwd()),
-)
+if (os.path.exists('1.cpp')):
+    print('1.ccp already exists')
+else:
+    shutil.copy(
+        '{0}/_lib/cpp/oj-configs/tpl-leet.cpp'.format(project_root),
+        '{0}/1.cpp'.format(os.getcwd()),
+    )
 run('code 1.cpp', shell=True, encoding='utf-8')
